[cff] Fix missing sign extension for charstring opcode 28
Opcode 28 encodes a signed 16-bit value, but the uint16_t result was
assigned to int32_t without sign extension, making negatives positive.

+++ src/cff_charstring.cc
@@ -195,7 +195,7 @@ bool ReadNextNumberFromCharString(ots::Buffer *char_string,
       return OTS_FAILURE();
     }
     result += v;
-    *out_number = result;
+    *out_number = static_cast<int16_t>(result);
   } else if (v <= 31) {
     *out_number = v;
     *out_is_operator = true;
